homepagecalculationfunctionworking
